Move files from `types/transactions` to `types` directory
