+++ lib/src/main/kotlin/com/swmansion/starknet/account/Account.kt
@@ -2,7 +2,6 @@ package com.swmansion.starknet.account
 
 import com.swmansion.starknet.data.TypedData
 import com.swmansion.starknet.data.types.*
-import com.swmansion.starknet.data.types.transactions.*
 import com.swmansion.starknet.provider.Request
 
 /**
