+++ lib/src/main/kotlin/com/swmansion/starknet/data/types/Payloads.kt
@@ -1,6 +1,5 @@
 package com.swmansion.starknet.data.types
 
-import com.swmansion.starknet.data.types.transactions.*
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
