+++ lib/src/main/kotlin/com/swmansion/starknet/data/types/Resources.kt
@@ -1,6 +1,6 @@
 @file:JvmName("Resources")
 
-package com.swmansion.starknet.data.types.transactions
+package com.swmansion.starknet.data.types
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
