+++ lib/src/main/kotlin/com/swmansion/starknet/data/types/SimulatedTransaction.kt
@@ -1,7 +1,6 @@
-package com.swmansion.starknet.data.types.transactions
+package com.swmansion.starknet.data.types
 
 import com.swmansion.starknet.data.serializers.TransactionTracePolymorphicSerializer
-import com.swmansion.starknet.data.types.*
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
