+++ lib/src/main/kotlin/com/swmansion/starknet/data/types/SimulatedTransactionList.kt
@@ -1,7 +1,6 @@
 package com.swmansion.starknet.data.types
 
 import com.swmansion.starknet.data.serializers.SimulatedTransactionListSerializer
-import com.swmansion.starknet.data.types.transactions.SimulatedTransaction
 import kotlinx.serialization.Serializable
 
 @JvmInline
