+++ lib/src/main/kotlin/com/swmansion/starknet/data/types/TransactionReceipt.kt
@@ -1,6 +1,5 @@
 package com.swmansion.starknet.data.types
 
-import com.swmansion.starknet.data.types.transactions.ExecutionResources
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
