+++ lib/src/main/kotlin/com/swmansion/starknet/provider/Provider.kt
@@ -1,7 +1,6 @@
 package com.swmansion.starknet.provider
 
 import com.swmansion.starknet.data.types.*
-import com.swmansion.starknet.data.types.transactions.*
 import com.swmansion.starknet.provider.exceptions.RequestFailedException
 
 /**
