+++ lib/src/test/kotlin/network/account/AccountTest.kt
@@ -4,7 +4,6 @@ import com.swmansion.starknet.account.StandardAccount
 import com.swmansion.starknet.crypto.StarknetCurve
 import com.swmansion.starknet.data.ContractAddressCalculator
 import com.swmansion.starknet.data.types.*
-import com.swmansion.starknet.data.types.transactions.*
 import com.swmansion.starknet.deployercontract.StandardDeployer
 import com.swmansion.starknet.extensions.toFelt
 import com.swmansion.starknet.extensions.toUint256
